requirements.txt
aiogram==3.22.0
openai==1.102.0
python-dotenv==1.1.1